Fix: - Remove unnecessary imports from effects -Rename logger module
