+++ effects/antidot_effect.py
@@ -2,7 +2,7 @@ import message_text_config as msg
 import utils
 from citizens.citizen import Citizen
 from citizens.player import Player
-from murder_logging import logger
+
 
 from effects.effect import Effect, InputStatusCode
 
@@ -13,7 +13,6 @@ class AntidotEffect(Effect):
     def __init__(self, context: 'Context', name: str,
                  creator: Citizen) -> None:
         super().__init__(context, name, creator)
-        self.logger = logger.getChild(__name__)
 
     def _activate_impl(self) -> bool:
         target_number = utils.read_target_number(
