+++ effects/gang_effect.py
@@ -2,7 +2,7 @@ import message_text_config as msg
 import utils
 from citizens.citizen import Citizen
 from citizens.player import Player
-from murder_logging import logger
+
 
 from effects.effect import Effect, EffectStatus, InputStatusCode
 
@@ -11,7 +11,6 @@ class GangEffect(Effect):
     def __init__(self, context: 'Context', name: str,
                  creator: Citizen) -> None:
         super().__init__(context, name, creator)
-        self.logger = logger.getChild(__name__)
 
     def _activate_impl(self) -> bool:
         self.targets.append(self.city)
