+++ effects/gossips_effect.py
@@ -4,7 +4,7 @@ from typing import List
 import message_text_config as msg
 import utils
 from citizens.citizen import Citizen
-from murder_logging import logger
+
 
 from effects.effect import Effect, InputStatusCode
 
@@ -14,7 +14,6 @@ class GossipsEffect(Effect):
                  creator: Citizen) -> None:
         super().__init__(context, name, creator)
         self.night_number = 0
-        self.logger = logger.getChild(__name__)
 
     def _activate_impl(self) -> bool:
         target_number = utils.read_target_number(
