+++ effects/spy_effect.py
@@ -3,7 +3,7 @@ from typing import Tuple
 import message_text_config as msg
 import utils
 from citizens.citizen import Citizen
-from murder_logging import logger
+
 
 from effects.effect import Effect, InputStatusCode
 
@@ -14,7 +14,6 @@ class SpyEffect(Effect):
     def __init__(self, context: 'Context', name: str,
                  creator: Citizen) -> None:
         super().__init__(context, name, creator)
-        self.logger = logger.getChild(__name__)
 
     def _activate_impl(self) -> bool:
         self.targets.append(self.city.spy)
