+++ effects/theatre_effect.py
@@ -1,7 +1,7 @@
 import message_text_config as msg
 import utils
 from citizens.citizen import Citizen
-from murder_logging import logger
+
 
 from effects.effect import Effect, InputStatusCode
 from effects.staging_effect import StagingEffect
